+++ c/include/nnstreamer-tizen-internal.h
@@ -13,6 +13,7 @@
 #define __TIZEN_MACHINELEARNING_NNSTREAMER_INTERNAL_H__
 
 #include <nnstreamer.h>
+#include <nnstreamer-single.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -24,6 +25,24 @@ extern "C" {
  * @since_tizen 5.5
  */
 int ml_pipeline_construct_internal (const char *pipeline_description, ml_pipeline_state_cb cb, void *user_data, ml_pipeline_h *pipe);
+
+/**
+ * @brief An information to create single-shot instance.
+ */
+typedef struct {
+  ml_tensors_info_h input_info;  /**< The input tensors information. */
+  ml_tensors_info_h output_info; /**< The output tensors information. */
+  ml_nnfw_type_e nnfw;           /**< The neural network framework. */
+  ml_nnfw_hw_e hw;               /**< The type of hardware resource. */
+  char *models;                  /**< Comma separated neural network model files. */
+  char *custom_option;           /**< Custom option string for neural network framework. */
+} ml_single_preset;
+
+/**
+ * @brief Opens an ML model with the custom options and returns the instance as a handle.
+ * This is internal function to handle various options in public APIs.
+ */
+int ml_single_open_custom (ml_single_h *single, ml_single_preset *info);
 
 #ifdef __cplusplus
 }
