+++ tests/capi/unittest_capi_inference_nnfw_runtime.cc
@@ -11,7 +11,8 @@
 #include <glib.h>
 #include <nnstreamer.h>
 #include <nnstreamer-single.h>
-#include <nnstreamer-capi-private.h>
+#include <ml-api-internal.h>
+#include <ml-api-inference-internal.h>
 
 static GMutex g_test_mutex;
